MAINT: cleanup circular import b/w arrayprint.py,numeric.py
Moves array_repr, array_str, g/set_printoptions to arrayprint.py
